refactor: 统一 utils 导出入口并清理未使用的导出函数，减少死代码和维护负担
