+++ src/renderer/messages/MessageList.tsx
@@ -18,7 +18,7 @@ import classNames from 'classnames';
 import React, { createContext, useEffect, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Virtuoso } from 'react-virtuoso';
-import { uuid } from '../utils/common';
+import { uuid } from '../utils';
 import HOC from '../utils/HOC';
 import MessageCodexToolCall from './codex/MessageCodexToolCall';
 import type { FileChangeInfo } from './codex/MessageFileChanges';
