+++ src/renderer/utils/shareNotify.tsx
@@ -23,10 +23,7 @@ export function showShareLoading(options?: ShareNotifyOptions): string {
   Notification.info({
     id: SHARE_NOTIFY_ID,
     title: options?.title || 'ShareOne',
-    content: React.createElement('div', { className: 'flex items-center gap-8px' },
-      React.createElement('span', { className: 'arco-notification-loading-icon' }),
-      React.createElement('span', null, '分享中...')
-    ),
+    content: React.createElement('div', { className: 'flex items-center gap-8px' }, React.createElement('span', { className: 'arco-notification-loading-icon' }), React.createElement('span', null, '分享中...')),
     duration: 0, // Don't auto close
     closable: false,
     position: 'topRight',
@@ -45,7 +42,7 @@ export function updateShareSuccess(id: string, url: string): void {
   });
 
   const handleCopy = () => {
-    copyText(url);
+    void copyText(url);
     Notification.success({
       id: 'shareone-copy-success',
       title: 'ShareOne',
@@ -58,22 +55,30 @@ export function updateShareSuccess(id: string, url: string): void {
   Notification.success({
     id: id || SHARE_NOTIFY_ID,
     title: 'ShareOne',
-    content: React.createElement('div', { className: 'flex flex-col gap-8px' },
-      React.createElement('div', { className: 'flex items-center gap-8px' },
-        React.createElement('span', { className: 'text-13px color-success font-500' }, '分享成功，链接已复制')
-      ),
-      React.createElement('div', { className: 'flex items-center gap-8px' },
-        React.createElement('a', {
-          href: url,
-          target: '_blank',
-          rel: 'noopener noreferrer',
-          className: 'text-12px color-primary underline break-all max-w-200px truncate hover:opacity-80'
-        }, url),
-        React.createElement('span', {
-          className: 'cursor-pointer color-primary opacity-70 hover:opacity-100 transition-opacity',
-          onClick: handleCopy,
-          style: { lineHeight: 0 }
-        },
+    content: React.createElement(
+      'div',
+      { className: 'flex flex-col gap-8px' },
+      React.createElement('div', { className: 'flex items-center gap-8px' }, React.createElement('span', { className: 'text-13px color-success font-500' }, '分享成功，链接已复制')),
+      React.createElement(
+        'div',
+        { className: 'flex items-center gap-8px' },
+        React.createElement(
+          'a',
+          {
+            href: url,
+            target: '_blank',
+            rel: 'noopener noreferrer',
+            className: 'text-12px color-primary underline break-all max-w-200px truncate hover:opacity-80',
+          },
+          url
+        ),
+        React.createElement(
+          'span',
+          {
+            className: 'cursor-pointer color-primary opacity-70 hover:opacity-100 transition-opacity',
+            onClick: handleCopy,
+            style: { lineHeight: 0 },
+          },
           React.createElement(Copy, { theme: 'outline', size: '14' })
         )
       )
@@ -91,19 +96,9 @@ export function updateShareError(id: string, errorMsg: string): void {
   Notification.error({
     id: id || SHARE_NOTIFY_ID,
     title: 'ShareOne',
-    content: React.createElement('div', { className: 'flex flex-col gap-4px' },
-      React.createElement('span', { className: 'text-13px color-error font-500' }, '分享失败'),
-      React.createElement('span', { className: 'text-12px color-text-3 break-all max-w-200px' }, errorMsg)
-    ),
+    content: React.createElement('div', { className: 'flex flex-col gap-4px' }, React.createElement('span', { className: 'text-13px color-error font-500' }, '分享失败'), React.createElement('span', { className: 'text-12px color-text-3 break-all max-w-200px' }, errorMsg)),
     duration: 8000,
     closable: true,
     position: 'topRight',
   });
-}
-
-/**
- * Close share notification
- */
-export function closeShareNotify(id?: string): void {
-  Notification.remove(id || SHARE_NOTIFY_ID);
 }
